the rest of service and controller

+++ src/main/java/edu/pja/sii_lat/service/IExchangeRateService.java
@@ -1,11 +1,10 @@
 package edu.pja.sii_lat.service;
 
-import java.math.BigDecimal;
 import java.util.Set;
 
 public interface IExchangeRateService {
 
-    BigDecimal getRateFor(String currencyCodeFrom, String currencyCodeTo);
+    double getRateFor(String currencyCodeFrom, String currencyCodeTo);
 
     Set<String> getValidCurrencies();
 }
